flask-app/requirements.txt
APScheduler==3.10.4
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Session==0.8.0
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msal==1.31.0
msgspec==0.18.6
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.5
WTForms==3.2.1
